+++ src/main/java/com/oroarmor/multiitemlib/mixin/CapeFeatureRendererMixin.java
@@ -1,0 +1,19 @@
+package com.oroarmor.multiitemlib.mixin;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.oroarmor.multiitemlib.api.UniqueItemRegistry;
+import net.minecraft.client.render.entity.feature.CapeFeatureRenderer;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(CapeFeatureRenderer.class)
+public class CapeFeatureRendererMixin {
+    @WrapOperation(method = "render(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;ILnet/minecraft/client/network/AbstractClientPlayerEntity;FFFFFF)V",
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+    private boolean isAcceptableItem(ItemStack instance, Item item, Operation<Boolean> original) {
+        return UniqueItemRegistry.ELYTRA.isItemInRegistry(instance.getItem());
+    }
+}