+++ src/main/java/com/oroarmor/multiitemlib/mixin/LivingEntityMixin.java
@@ -1,0 +1,18 @@
+package com.oroarmor.multiitemlib.mixin;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.oroarmor.multiitemlib.api.UniqueItemRegistry;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(LivingEntity.class)
+public class LivingEntityMixin {
+    @WrapOperation(method = "tickFallFlying", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+private boolean handleDisableShield(ItemStack instance, Item item, Operation<Boolean> original) {
+    return UniqueItemRegistry.ELYTRA.isItemInRegistry(instance.getItem());
+}
+}